requirements.txt
# ─── Core Framework ───────────────────────────────────────────
langchain==0.3.14
langchain-community==0.3.14
langchain-huggingface==0.1.2

# ─── PDF Parsing ──────────────────────────────────────────────
pymupdf==1.25.1
fitz 

# ─── Embeddings ───────────────────────────────────────────────
sentence-transformers==3.3.1

# ─── Vector Store ─────────────────────────────────────────────
faiss-cpu==1.9.0.post1

# ─── LLM Access ───────────────────────────────────────────────
huggingface-hub==0.27.1

# ─── Config & Environment ─────────────────────────────────────
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1

# ─── UI ───────────────────────────────────────────────────────
streamlit==1.41.1

# ─── Dev Tooling ──────────────────────────────────────────────
black==24.10.0
ruff==0.9.2
pytest==8.3.4